LPS-99657 Move classes to 'internal'
