+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/security/permission/JournalFeedPermissionUpdateHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.security.permission;
+package com.liferay.journal.internal.security.permission;
 
 import com.liferay.journal.model.JournalFeed;
 import com.liferay.journal.service.JournalFeedLocalService;
