+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/security/permission/JournalFolderPermission.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.security.permission;
+package com.liferay.journal.internal.security.permission;
 
 import com.liferay.journal.model.JournalFolder;
 import com.liferay.journal.service.JournalFolderLocalService;
