+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/security/permission/JournalPermission.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.security.permission;
+package com.liferay.journal.internal.security.permission;
 
 import com.liferay.journal.constants.JournalConstants;
 import com.liferay.portal.kernel.exception.PortalException;
