+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/trash/JournalArticleTrashHandler.java
@@ -18,13 +18,13 @@ import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
+import com.liferay.journal.internal.security.permission.JournalArticlePermission;
+import com.liferay.journal.internal.security.permission.JournalFolderPermission;
 import com.liferay.journal.internal.util.JournalHelperUtil;
 import com.liferay.journal.internal.util.JournalUtil;
 import com.liferay.journal.model.JournalArticle;
 import com.liferay.journal.model.JournalArticleResource;
 import com.liferay.journal.model.JournalFolder;
-import com.liferay.journal.security.permission.JournalArticlePermission;
-import com.liferay.journal.security.permission.JournalFolderPermission;
 import com.liferay.journal.service.JournalArticleLocalService;
 import com.liferay.journal.service.JournalArticleResourceLocalService;
 import com.liferay.journal.service.JournalFolderLocalService;
