+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/trash/JournalFolderTrashHandler.java
@@ -17,11 +17,11 @@ package com.liferay.journal.trash;
 import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.journal.exception.InvalidDDMStructureException;
+import com.liferay.journal.internal.security.permission.JournalFolderPermission;
 import com.liferay.journal.internal.util.JournalHelperUtil;
 import com.liferay.journal.internal.util.JournalUtil;
 import com.liferay.journal.model.JournalFolder;
 import com.liferay.journal.model.JournalFolderConstants;
-import com.liferay.journal.security.permission.JournalFolderPermission;
 import com.liferay.journal.service.JournalFolderLocalService;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.ContainerModel;
